// src/types/UpgratedProduct.ts
export type UpgratedProduct = {
  id: string;
  category: string;
  phoneId: string;
  itemId: string;
  name: string;
  fullPrice: number;
  price: number;
  screen: string;
  capacity: string;
  color: string;
  ram: string;
  year: number;
  image: string;
  addedToFavourites: boolean;
  addedToCart: boolean;
  quantity: number;
};